backend/requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
zarr>=2.16.0
numcodecs>=0.11.0
faiss-cpu>=1.7.4

# RAG pipeline
geopy>=2.4.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Visualization
folium>=0.16.0

# LLM (optional — set GEMINI_API_KEY to enable)
google-generativeai>=0.4.0

# Utilities
tqdm>=4.65.0
pyarrow>=14.0.0
python-multipart>=0.0.9
